+++ packages/react/src/Label/index.ts
@@ -1,4 +1,3 @@
 import Label from './Label'
 export type {LabelProps, LabelColorOptions} from './Label'
-export {variants} from './Label'
 export default Label