+++ packages/styled-react/src/experimental.tsx
@@ -1,7 +1,30 @@
+export {Dialog, type DialogProps} from './components/Dialog'
 export {
   PageHeader,
   type PageHeaderProps,
   type PageHeaderActionsProps,
   type PageHeaderTitleProps,
 } from './components/PageHeader'
-export {Dialog, Table, Tooltip, UnderlinePanels} from '@primer/react/experimental'
+
+export {Tooltip, type TooltipProps} from './components/Tooltip'
+
+export {
+  UnderlinePanels,
+  type UnderlinePanelsProps,
+  type UnderlinePanelsTabProps,
+  type UnderlinePanelsPanelProps,
+} from './components/UnderlinePanels'
+
+export {
+  Table,
+  type DataTableContainerProps,
+  type TableProps,
+  type TableHeadProps,
+  type TableBodyProps,
+  type TableRowProps,
+  type TableHeaderProps,
+  type TableCellProps,
+  type TableTitleProps,
+  type TableSubtitleProps,
+  type TableActionsProps,
+} from './components/DataTable'